BUILD: buffer: fix another isprint() warning on solaris
This one came with commit recent be0efd8. Solaris wants ints, not chars.

+++ src/buffer.c
@@ -217,7 +217,7 @@ void buffer_dump(FILE *o, struct buffer *b, int from, int to)
 		}
 		fprintf(o, "  ");
 		for (i = 0; (from + i < to) && (i < 16) ; i++) {
-			fprintf(o, "%c", isprint(b->data[from + i]) ? b->data[from + i] : '.') ;
+			fprintf(o, "%c", isprint((int)b->data[from + i]) ? b->data[from + i] : '.') ;
 			if ((((from + i) & 15) == 15) && ((from + i) != to-1))
 				fprintf(o, "\n");
 		}
